nintendo/dkong.cpp: add trainer ROM set
This hacked version of Donkey Kong is made to help “train” players on
the game as they learn the game mechanics and all the subtle details
involved with player and enemy object movement.

There was an older version 1.00 of the ROM set, but I do not believe
it was backed up by the WayBack Machine.  At least I was unable to
find it.

From: https://donkeykongforum.net/index.php?topic=1642.90