psikyo.cpp : Fixes an off-by-one error in m_screen->set_raw
Fixes a off-by-one error where number of visible lines were 223 instead of 224.

There are indeed 39 HSync PULSES between the visible color data... but the first one of those are signaling the end of the previous visible frame, and the last one of them is signaling the start of first visible row of the next frame, which means that there's 38 "non-visible" lines during blanking.

This means:
262 total lines. 224 visible. 38 blanking.