psikyo/psikyo.cpp: Fixed docs for visible area and vblank of Tengai
The code for Tengai was updated to be correct, but not the docs. This fixes the docs too.

I was looking at sync-pulses surrounding front/back-porch, and sync pulses during vblank previously. This is an incorrect way of looking at things.

See: http://img.buffis.com/psikyo/tengai/vblank.png
which has a very easy to grasp visual representation of the VBlank lines.