psikyo.cpp: Remove MACHINE_IMPERFECT_TIMING and fix some docs
- Remove MACHINE_IMPERFECT_TIMING from psikyo.cpp roms
- Minor documentation fixes (IRQ info, typo for LH5168)

Main motivation for the incorrect timing seems to be that these have four frames of input lag. This is actually consistent with PCBs though, unlike what the comment states.

When doing measurements with a 240fps iphone camera and "Is it snappy" on Tengai, the current behavior is:
- Input is triggered during a frame (and likely sampled during vblank?)
- Three frames of inaction follows
- The fourth frame afterwards, the sprite moves.

With sprites at top of screen, this gives delays of 54.2ms -> 62.5ms for me from input press, to sprite changing.

This is consistent with mame behavior:
- Press pause
- Hold an input and step-pause until movement
- Movement will trigger on fourth frame

Existing comment was added in:
https://github.com/mamedev/mame/commit/a9222ff5d03e1db73c866aab659e77a22718e208