video/epic12.cpp: Fix clipping of CV1K games
Change clipping for CV1K games to draw 32 pixels surrounding the visible area.

This can be easily seen in Muchi Muchi Pork, which has a VRAM viewer in Special mode (Object Test), which will show 32 px drawn around the visible areas of framebuffers.

For most gamers, this wont really matter at all... except for in Muchi Muchi Pork, where changing this actually fixes a bug for Rafute.

When Bombing with Rafute, the screen background will go wavy in a sine-like pattern. Currently in mame, the top of screen will show black pixels when this happens.

With this fix for clipping, the background will instead be visible correctly.

Also renamed the "scroll registers" to have it more clear which one of these are actually used as a "scroll register" (or rather offset for drawing), and which one is strictly used for clipping.