video/epic12: Apply clipping to Blitter calculations as well
Fix issue where some games do very large over-draws which when utilized in Blitter calculations will cause more delay than whats seen on hardware. The fog in S1 of Mushi Futari 1.5 is a good example of this.

On Mushi Futari 1.5, in the section before the boss, the Blitter delay seen on PCB is aprox 12 milliseconds at peaks. With this fix, it matches that quite well!

It's not 100% clear to me this is exactly how things behave, but this should be closer to the truth, until I can figure out a more reliable and reproducible test-case.

See discussions on this PR for background : https://github.com/mamedev/mame/pull/11227

For reference, Saleae logic analyzer output is available in:
http://img.buffis.com/misc/futaris.sal