namco54: fix interrupt handling
There seems to be a bit flip in the 54xx ROM dump. Byte 0x10 is 0x3E,
instruction EN #04, enabling interrupts. This occurs early in the
interrupt routine and has no effect as interrupts are already enabled.
At byte 0x174 is another EN #04, this time correct as it enables
interrupts right before ending the routine. This happens again at 0x1BD,
0x1D9, 0x1F6, and 0x214.

This first EN should be a DIS, disabling interrupts during the interrupt
routine. This is opcode 0x3F, so the low bit seemingly got flipped.