Revert "Clean up the Votrax SC-01[-A] class names and device names. [Lord Nightmare]"
Names in the global namespace need to be specific.  Making them overly
generic is a very bad idea, and leads to confusion and conflicts.
Device type names must always be in the global namespace, and most
device classes are in the global namespace.

Changing members like m_votrax to m_sc01 doesn't help understand the
code.  If there's only one Votrax speech module in the system, it's
obvious what it is.

This reverts commit c6f0e83ecfb97c298ffcf16e136496054a88b1ca.