samsho3j
updated to known source based on digital storefront, validated that the psx catalog/file list was identical in terms of size/timestamp, all files have identical hashes except the files in the 'xa' directory which should be audio - seems to point to a different dumping method/error correction - binary compare shows majority is identical and the changes are mostly in 'empty' areas on the new dump.  New CHD loads correctly.