Migrate cv1k driver to use nandflash.h instead of serflash.h
serflash.h looks real old and sortof messy, and is pretty much implementing the basic functionality of K9F1G08U0 (it has that manufacturer/device ID hardcoded).

It's only used by cv1k and nexus3d.cpp (which I'll fix in a followup).

nandflash.h seems newer, has support for SAMSUNG_K9F1G08U0B and looks less jank

Migrating to it is pretty straightforward and seems to work.

Only difference is the write to 0x10C00003 which controls the CE pin of U2 (through the U13 CPLD). serflash seems to have a setter for this (flash_enab_w), but I don't really see how it's needed. As far as I can understand, its only used to signal that U2 should be the one to access the shared data-bus, which is a hardware limitation not relevant for emulation.

I assume that's why there's no handling of CE pin in nandflash.h/cpp

Tried a few different games, and seems to work fine