strtdriv: fixed incorrect brake pedal port mapping in harddriv.cpp
Fixed wrongly mapped brake pedal input from 8ADC:6 to 8ADC:2 and swapped the Voice Mic function which is unused from 8ADC:2 to 8ADC:6 to avoid any conflict/loss of functionality. Steering input still needs to be fixed (if you turn to the far left of right the centre point of the steering wheel will shift and constantly pull the wheel in the opposite direction) Hard Drivin's Airborne has the same problem. Will try to fix that if I can and update the code here on GitHub.