Fix steering for ALL controller using cuavas code suggestion harddriv.cpp
It turns out my steering fix code didn't work for xbox one controllers but it did for my PS5 controller so I'm using the code suggested to me by cuavas to fix it for all controllers. I can only test it on Xbox One controller, PS5 and Logitech G923 wheel but if it works on them i assume it should work for others too.