More accurate input descriptions in the input assignment menu.cpp
Defined PORT_NAME for inputs that make more sense so the user knows what they do. 

Discovered if i add a PORT_NAME for "PORT_BIT( 0x4000, IP_ACTIVE_LOW, IPT_CUSTOM )  /* center edge on steering wheel */" i can assign a button to it in the input assignment menu that lets you reset the centre point of the steering wheel/joystick. Unfortunately it doesn't seem work when playing the game so sometimes the steering will still go out of alignment and cant be fixed unless you crash the car or run out of time and pressing this button whilst the wheel or joystick is centred doesn't help.  Was thinking about trying to find where the centre point is stored in the games memory and making a cheat file that locks it to the right value that could maybe be enabled by default? I should maybe try figuring out what the unused PORT_BIT's do.