Added back original comments for "PORT_BIT(0x4000" - harddriv.cpp
Tested this more and it is in fact a integral part of how the wheel operates just not how I expected, if you remove it the wheel won't work in game so it's got to be doing something even if it appears when you manually activate it that it's only being applied or used in the test menu. Was investigating it in a effort to try and fix or mitigate the steering wheel centring bug in Hard Drivin's Airborne but i think it's a game code issue and not a emulator issue and i've tried different ways to get around it using the emulator source code related to the functioning of this game and can't find a work around.