Set ADC0809 to 640Khz to match chips Datasheet, harddriv.cpp
Max clock for this chip is 640Khz and minimum is 500Khz. Don't know what the real hardware was set to so using the highest of the 2 clock speeds. Updated the note regarding it next to the source code.

It doesn't change how the emulator runs the games but will be more accurate than the 100Khz it was set to as a placeholder.