input/input_sdl.cpp: Added an SDL lightgun provider.
This does essentially the same thing as the Win32 lightgun provider,
mapping the absolute pointer position over the window to gun axes.

Also added a bunch of const in the windows input handling code.

docs: Bumped version, as features that are not in a releaesd version of
MAME are now documented.