Fixed loading of collision detection files in "hdrivairp". "Airborne" track now playable!
I fixed how the sboota.bin gets loaded into system memory swapped it from "dsp32" to "asic65" and changed read format of the collision detection files roads.0, roads.1 etc from "rom_load16_byte" to "rom_load32_byte" so now the game works like it should once you actually get it to load, takes many attempts for the game to load up a level because of "bad poly buff error" but once your in game it is now playable and is a more complete version of the game than "hdrivair" it has extra tracks and the voice recording function is implemented. Extra buildings and objects in the Mountain track too. Been trying to figure out how to get the game to load up properly without trying 20 times but no luck yet but this is a major step forward. Has the test menu brake calibration fix for harddrivc and racedrivc too.