emucore.h: fix build error on VC++
This compiler requires exception classes to be copyable, because it copies
the exception object as part of `std::current_exception()`. This means that
code like the following has to work:

    emu_fatalerror e1(...);
    emu_fatalerror e2(e1);

However, it doesn't, because in this case the perfect-forwarding constructor
is a better match than the copy constructor. So the compiler tries to
instantiate the perfect-forwarding constructor with `Format`=`emu_fatalerror &`,
which fails, since that's not a valid format string type.

To fix this, disable the perfect-forwarding constructor when `Format` is a
subclass of `emu_fatalerror`.