imagedev/floppy: fix a bug with reading the first 1-bit on a track (#12439)
`floppy_image_device::find_index` uses binary search to find the index for
which `buf[spos] <= position < buf[spos + 1]`. However, the algorithm behaves
incorrectly when `position < buf[0]`. In this case, the algorithm returns 0,
as if `position` was between `buf[0]` and `buf[1]`.

The effect of this is that if `get_next_transition` is called with a timestamp
that is between the start of the revolution and the first transition, then
instead of returning the timestamp of that transition, it returns the timestamp
of the second transition instead. Essentially, the first 1-bit on the track gets
flipped to a 0.

I have encountered this in Apple II emulation, where this bug manifests as
sporadic I/O errors.

Fix it by doing two things:

1. Replace `find_index` with a call to `upper_bound` from the standard library,
   which behaves correctly in edge cases.

2. If `upper_bound` signals that `position < buf[0]`, then adjust `base` and
   `index` to point to the last transition of the previous revolution.