upd765: Drive busy bits should not be cleared until a SIS is executed
The upd765 does not clear the drive busy bits until you issue a sense
interrupt status command. Without this fix CP/M Plus on the QX-10 gives
disk I/O errors because it expects the busy bit to stay set until it has
a chance to issue the SIS command.