Fix MAME Debugger GUI issues
1) horizontal disassembly scrolling yielded garbage characters in view

debug_view_disasm::redraw was passing to print a relative column number (dependent on where the horizontal scrollbar was scrolled) for the "end" parameter when printing the rightmost chunk of text for each line.  Changed to an absolute column number (independent of where the horizontal scrollbar was scrolled), as expected by the print function.

Changed some int types and var naming to be consistent.

2) The parent debugger window refused to be widened or horizontally scrolled even when disassembly contents very wide (e.g., due to comments).

debug_view_disasm: Replace hard-coded value of 50 for comment length with a dynamically calculated value based on the comments retrieved for the current buffer of disassembly to be shown (debug_view_disasm::complete_information).

consolewin_info: Factor code that computes minwidth/maxwidth based on child windows so it may also be called from consolewin_info::recompute_children.

debugwin_info::window_proc: Added call to recompute_children() at top of case WM_GETMINMAXINFO.  This ensures that when the system queries the parent window's min/max width, the parent first consults its children (specifically the disassembly window) so it calculates the values properly.