MS68K: Marion Systems MS68K
The Marion Systems MS68K was a single-board 68000 system designed by Tom
Oberheim in 1985 with a DUART providing two serial ports and (in combination
with a latch) a parallel printer port, WD1772 floppy controller, optional
NCR53C80 SCSI controller, and up to 512KB of RAM and 128KB of ROM onboard.
Also included is an expansion connector providing access to most of the 68000
bus as well as memory refresh signals and a single dedicated interrupt line.

The stock ROM is a monitor that also contains a port of Tiny BASIC 1.2 at
0xF04900 with a loader at 0xF04800. Thus while there is no direct monitor
command to start it, a simple GF04800 will start Tiny BASIC (and as usual BYE
will exit back to the monitor).

The monitor’s included floppy boot functionality will load the first sector of
the disk to 0x800 and jump to it in supervisor mode. There is no SCSI boot
functionality built in.

The MS68K was used in a number of college and university computer systems
classes, and a variant of Peter Stark’s SK*DOS was available.