hash/ibm5170{,_cdrom}.xml: Pristine Windows NT 3.1 builds
The image that was listed here is a hacked-up warez version from 2003
that doesn't function properly (won't install on most computers NT 3.1
should be able to install on).

In its stead, this is a list of pristine images direct out of
Microsoft, no modifications have been made to them, and NT 3.1's
(admittedly limited) full hardware compatibility is supported.

Taking from the BeOS example, floppy images that are intimately tied
to the CD-ROM are listed in the same entry in ibm5170_cdrom.xml, since
they are only useful in booting a computer in order to install from
CD-ROM (which requires a compatible SCSI controller and SCSI CD-ROM
drive).  Installing from MS-DOS is always another possibility.