hash/ibm5170_cdrom: add/fix Windows 2000
The Windows 2000 image existing in prior commits was _slightly_ off an
official build.  Official builds use the AutoCRC feature of cdimage to
make sure images have a CRC32 hash of 0xffffffff, accomplished by
editing a few bytes in extraneous sectors of the ISO-9660 file system
that are not user visible.  The prior image did not have these bytes,
albeit everything else was a match.

Listed here are a complete set of Windows 2000 Professional, Server,
Advanced Server in all service pack levels and RTM.  The "Select"
licensing disc was used for all of them; this version was released for
volume licensing and does not require a product key to be entered.

Also included are the MUI language packs and standalone service pack
1-4 installers, sourced from official CD-ROMs also.