addrmap: Allow ioport finders in portr/w/rw
Convert a number of drivers to use ioport finders in the memory map.

This is not entirely gratuitous.  The idea is to ensure all the ports
are in the class, so that eventually ioport finders become ioport
creators.  The ultimate aim is to C++-ify port descriptions which
would be made way nicer by getting rid of the need for PORT_START.