Game Boy: add new BIOS variants and gbplayer (not working GBA clone)
These files are all floating around the No-Intro sets.

For gbcolor, a Japan-exclusive version of the BIOS is added.  The
variant of the BIOS from the Game Boy Advance is also included, which
unlocks features in a few games that use it to detect running on GBA
hardware.

A beta version of the GBA BIOS has been added, which features a
different boot up sound and lacks the ability to escape to network
boot mode (start+select in the released console).

The Game Boy Player has been added as a GBA clone.  At least the BIOS
checksum is documented here, but it fails to boot in the MAME gba
driver.