debugger/osx: Don't use uninitialised NSRect when creating pop-up button.
The initial size doesn't really matter because the pop-up button is
resized before being added to the container that holds it and the
expression field.  The disassembly viewer apparently wasn't updated
properly like the memory viewer when the auto-sizing code was
refactored.

Hopefully this will fix the debugger crashing when restoring window
positions from configuration (although I would have thought this would
cause crashes in more situations if it was the cause of that).