ap2_dsk.cpp: fix loading bardstl3 (#13055)
I accidentally broke this in 47d8c69b, where I forgot to account for the
"oddball" images from the software list. AFAICS, bardstl3 is the only software
to be broken, because its images are _shorter_ than they're supposed to be.

In addition, disk 3 of this image has a different size that's not listed in
`identify()`, so it couldn't even be loaded before that commit. Fix that too.