vector_device: render dots as fuzzy circles.
When emulating games like Asteroids with increased -beam_dot_size,
MAME's various renderers draw the dots differently:

rendersw (gdi and libretro) draws a vertical line
bgfx draws a diamond
d3d draws a diamond
opengl draws a circle (reasonable)

This change replaces add_line() with add_quad() and a fuzzy circle
texture, so that dots look the same on all renderers.

Also increase the brightness of dots by 2X (capped to 255) because I
once played Asteroids on a real cabinet and found the dots surprisingly
bright. The Asteroids high score screen uses intensity 112 for the words
and 64 for the dots, which suggests that a factor of 2 is approximately
correct.

Related to https://github.com/mamedev/mame/issues/399