Revert "machine/ncr5385.cpp: clear status bits first when updating so they aren't stuck on (#13109)"
This reverts commit ff6d52d56a9dcac133e5fee805a277bf0f9b60a4.

Neither of the changes make sense.  The addition to
ncr5385_device::update_int is redundant is the bits are cleared just
above the if statement.  The addition to ncr5385_device::aux_status_r is
unnecessary as the bits are only set when an interrupt is asserted (if
they're set when it isn't, there's a logic bug elsewhere).