CD-i: Fix DYUV Format
This is part of a larger change, and has been broken out to reduce the complexity.

This change fixes the following issues with the DYUV format:

1. The DYUV format incorrectly was too dark by 16/256. This fix corrects a previously incorrect attempt to fix that issue.
2. The DYUV format does not allow Mosaic features.
3. The Transparency flag is left shifted incorrectly.

To Validate this change:
1. Use the Validation CD (Europe).
2. Click "Philips Tests"
3. Click "Video Data"
4. Click "DYUV Decoder"

Result:
Notice that the first two images are displayed in black and white. Previously, these images had random noise/color. The orderly pattern indicates the correct behavior.

Uncertain Issues:
The 3rd and 4th image in the test show pinkish/gray bars at the top. It is not clear whether it is correct or not. However this is not a regression.

The previous midpoint interpolation of 129+129 would be 1. The spec agrees with that formula. As the expected behavior is not yet well understood, a modified algorithm is used to produce more sensible midpoints that do not have this rollover artifact.

Known Issues:
There are further transparency issues that are out of scope for this update.

There are further format issues.