Revert transparency shift on DYUV
Reverts removal of transparency shift. Mattes must always run at double resolution, but here the image is generated at Normal resolution and then up-resed at the end of the function. Reverting this part of the change to remove the regression.

To test:
1. Run Guide to Heavy Duty Pumps in the Construction Market (USA)
2. When you mouse over the buttons on the title card, they should highlight normally.

The regression caused the highlight to be stretched horizontally to twice the expected width. That is now fixed.