Remove “Norton” prefix from PartitionMagic description
It’s a bit unusual in MAME to fully describe titles with a company
name prefixing it, except in cases where that name is intimately tied
to the product name.  That exception doesn’t apply here; this was
simply the version after Symantec purchased PowerQuest and it was no
longer “PowerQuest PartitionMagic”... at least until Symantec
proceeded to discontinue it entirely.