MC68EZ328: Allow selection and use of Port A
On EZ and VZ, PASEL is reserved, and it was hard-coded to 0 (disabled).
However, the user manual notes: "In an 8-bit-only system, you can
configure these pins as a parallel I/O port by writing a 1 to the
WDTH8 bit of the SCR."

According to the MC68328 datasheet, something similar should be needed
to activate its Port B, but that change is not included here. Its PBSEL
is exposed, so there is likely no difference in emulation.