ipc: Correct ROM sets for IPB
The ROM set used before is a modded version of v1.3 that runs uart1 at 9600 baud by default.
It's not clear if an IPB ever shipped with a v1.3 monitor from the factory, although many IPBs were upgraded to this monitor version by their operators, so it's better to default to v1.2 - either way, reference the real, genuine ROMs. not some random modified one.