util/chd.cpp: Hacked around GitHub #13029 (issues creating delta CHDs).
The code depends on getting away with trying to read at least one hunk
past the end of the parent CHD.  Previously it just ignored any errors
from reading the parent CHD, but adding error checking broke it.

This hack just ignores that specific error.  Also cleaned up some gross
flow control and made more things const.

The code should obviously be fixed properly, but that's non-trivial.
This gets it to a state no worse than it was in 0.270.