ipc: add alternate v1.2 ROM set found in an actual IPB
This ROM set was found in an IPB in my possession, programmed into a pair of Hitachi HN462716G EPROMs with 0J3 (September 1980) date code. The IPB in question has Siemens branding, which suggests that it was once part of a Siemens SME system (German OEM version of the Intellec Series II Microcomputer Development System AKA "imds2").

It's identical to the regular v1.2 in every way, except for serial channel A (TTY), which operates at 600 baud instead of the usual 110, and has an additional 400ms wait added after each line break. This seems to be an adaptation to a 600 baud teleprinter, perhaps a Telefunken Telestar 121/122.