act/victor9k.cpp: Get audio working on Victor 9000 (victor9k) (#13549)
* 6522via: Don't generate a signal on PB7 when the timer 1 latch is
programmed with 0.

On the Victor 9000, the clock for the audio codec is generated by a
via6522.  PB7 is connected to the codec's clock input.  Non-speech
sounds (e.g. beeps, musical notes) are produced by outputting a cyclical
waveform from the codec, with the pitch determined by the frequency of
the codec clock.

Software running on the Victor 9000 has been observed to attempt to
silence the audio by writing 0 to the T1 timer latch (not by turning off
continuous mode or PB7 output).  With the emulated via6522, this resulted
in a high-frequency clock signal being output on PB7, causing a
high-pitched squeal whenever notes are not being played.

From this observed behavior we could infer that the original 6522 HW
does not output a signal on PB7 when the latch value is 0, and that the
Victor 9000 software was relying on this behaviour to silence the audio
output (one would have to assume that the real hardware didn't produce
a squeal).

* mc6852: fix various transmit-related issues which were preventing
victor9k Audio output from working.

The expected behavior of mc6852 is to write the data received in the
FIFO register serially via tx_data_callback().  However, this was not
implemented, and the only way data in the transmit FIFO could be
removed and transmitted was by calling get_tx_byte(), and to then
serially transmit the data by some external mechanism.

Only m68sfdc.cpp calls get_tx_byte(), so it's hard to see how any
other device using the mc6852 would have been able to transmit data.

Software running on victor9k attempting to play audio would hang,
since it would block waiting for the TX FIFO to empty.  With these
changes, Victor 9000 audio playback works as expected, with the data
getting sent serially to the Audio codec.

In order to avoid breaking m68sfdc, a new API call is added to allow
data to be sent the "old" way.  m68sfdc now calls
set_tx_pull_mode(true), to get the previous behavior.  As I understand
it, other devices using mc6852 must be broken, and would need to
be revisited.

* victor9k: complete Audio support, adding a low-pass filter after
the HC-55516 codec.  Mark victor9k Audio as working.