cpu/drcbearm64.cpp: Fixed more cases where ROLAND clobbered source registers.
ROLAND could clobber source registers or produce incorrect results if
the desination was the same UML register as the shift or mask.

Fixed a bug in the logic for deciding whether to use a temporary
register for the mask for ROLINS.

Also optimised ROLAND and ROLINS some more (particularly translation
from UML's left rotate to ARM's right rotate) and removed a special case
that the simplifier now takes care of.