3rdparty/bgfx: Updated ImGui to 1.91.9b.
This was not straightforward, and we're now ahead of the ImGui version
in upstream bgfx.

Updated the bgfx-specific memory editor widget rather than just
partially removing the code.

Had to add an explicit template instantiation to imgui_widgets.cpp to
avoid a link error in the bgfx-specific range slider widget.

Got the imgconfig.h header a lot closer to the template from upstream
ImGui.