fix for issue 13685
Memory window address field: If address string evaluation fails, we'd fall back to the last-known-working string.  This works great if the failure is due to the user changing the string.  But if the user changes the source, that also can cause failures (different symbol table used for evaluation), and this fallback does not help.

The fix is to add a second fallback: Use "0" as the address string when all else fails.