New machines added ------------------ Block Fever [Dilweed, smf]
Sound frequencies are estimated from a video of the real game, as schematics do not exist.
The clickable artwork colors were taken from the same video.
An output is pulsed when you finish a game after scoring enough points to replay, it's purpose is unknown.
Better pictures of the PCB are required to confirm the layout.