igs/igs011.cpp: Added range check on blit destination.
Just log a message if a game tries to blit to an out-of-range layer for
the current configuration.

Also set vertical blanking time and total lines based on measurements
for nkishusp, and always blit to the whole 512*256 area as in principle
the CPU can read areas that aren't visible on-screen.