k007121: fix missing/extra sprites
I was recently doing some testing on contra hardware and found that
the 007121 IC has a 264 limit on the number of 8x8 sprite blocks it
will render.  It will either stop or start over once it hits the limit.

A number of games use this fact and pad a number of bugus/offscreen
32x32 sprites after their active sprites instead fully cleaning up
old sprite data.

We can use this limit to determine the number of sprites that should
be rendered.  This fixes missing sprites on combatsc, extra sprites
on contra, and maybe more.

An additional note from testing. If the 8x8 sprite block count was
at 262 and the game wanted to render a 16x16 sprite, the 007121 would
only draw 2 of the 4 8x8 sprite blocks.